create dill folder with build_program_files script if it does not exist, modify clear_dill_files function to check for dill folder before executing, add proposal name argument test and exception messages to compute_measures and join_inactives scripts

+++ seniority_list/functions.py
@@ -3221,9 +3221,10 @@ def clear_dill_files():
     used when changing case study, avoids possibility of file
     from previos calculations being used by new study
     '''
-    filelist = [f for f in os.listdir("dill/") if f.endswith('.pkl')]
-    for f in filelist:
-        os.remove('dill/' + f)
+    if os.path.isdir('dill/'):
+        filelist = [f for f in os.listdir('dill/') if f.endswith('.pkl')]
+        for f in filelist:
+            os.remove('dill/' + f)
 
 
 def load_datasets(other_datasets=['standalone', 'skeleton', 'edit', 'hybrid']):
